# src-tauri/docs/openapi.yml
openapi: "3.0.2"
info:
  title: API Title
  version: "1.0"
servers:
  - url: http://localhost:12321
paths:
  /association:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model_Association"
      tags:
        - Association
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model_Association"
      tags:
        - Association
  /association/{id}:
    get:
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model_Association"
      tags:
        - Association
    put:
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model_Association"
      tags:
        - Association
    delete:
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model_Association"
      tags:
        - Association
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int32
  schemas:
    Type_ID:
      type: integer
      format: int32
...
    Type_Date:
      type: string
      format: date-time
    Model_Association:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Type_ID"
        created_at:
          $ref: "#/components/schemas/Type_Date"
        updated_at:
          $ref: "#/components/schemas/Type_Date"
        name:
          type: string
      required:
        - id
        - created_at
        - updated_at
        - name
...